% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorFCI.R
\name{anchorFCI}
\alias{anchorFCI}
\title{anchorFCI: Causal Discovery using Anchors}
\usage{
anchorFCI(
  suffStat,
  indepTest,
  var_names,
  var_labels,
  prec_var_names = c(),
  prec_var_labels = c(),
  select_anchors = TRUE,
  m.max = 2,
  alpha = 0.05,
  fileid = "fit",
  conservative = FALSE,
  maj.rule = TRUE,
  renderAll = FALSE,
  width = 800,
  height = 450,
  output_folder = "./",
  verbose = TRUE
)
}
\arguments{
\item{suffStat}{A list of parameters used by the `indepTest` function.}

\item{indepTest}{A function that performs conditional independence tests and returns a p-value.}

\item{var_names}{Names of all variables to be represented as nodes in the graph, corresponding to dataset columns.}

\item{var_labels}{Labels used for `var_names` when generating plots.}

\item{prec_var_names}{Names of additional variables, where `prec_var_names` precede `var_names` in the partial ordering.}

\item{prec_var_labels}{Labels used for `prec_var_names` when generating plots.}

\item{select_anchors}{A Boolean indicating whether only a subset of `prec_var_labels` deemed reliable anchors should be used in the learning process.}

\item{m.max}{Maximum number of variables to condition on in conditional independence tests.}

\item{alpha}{Significance level for the conditional independence tests.}

\item{fileid}{A unique identifier string added to all saved objects.}

\item{conservative}{A Boolean indicating if the conservative approach should be applied.}

\item{maj.rule}{A Boolean indicating if the majority rule should be applied.}

\item{renderAll}{A Boolean indicating if all relevant plots should be rendered and saved within a subfolder in `output_folder`.}

\item{width}{Width of the rendered figures.}

\item{height}{Height of the rendered figures.}

\item{output_folder}{Path to the folder where output files should be saved.}
}
\description{
anchorFCI: Causal Discovery using Anchors
}
